Add class-validator and update TextToPdfService
Two new dependencies, 'class-validator' and 'class-transformer', have been added to the 'package.json' to enhance data validation. Also, the TextToPdfService has been refactored to only handle conversion of text to PDF. This change simplifies its responsibilities and makes the service more accurate in terms of Single Responsibility Principle (SRP). An appropriate validation decorator has been added to the 'create-text-to-pdf.dto.ts' to ensure the entered text is both present and of type 'string'.