Enable CORS and update Puppeteer launch method
CORS has been enabled in the main bootstrapping function to allow cross-origin requests. The Puppeteer launch method in the text-to-pdf service now includes '--no-sandbox' as an argument to enable running in non-sandboxed environments. Also, the way of setting response headers in text-to-pdf controller has been updated to use the 'set' method.